21.12.21 C, crawling update
21.12.21 C, crawling update

+++ team_project/src/model/coin/CoinVO.java
@@ -5,8 +5,8 @@ public class CoinVO {
 	private int coinpricekor;
 	private int coinpriceeng;
 	private int coinrate;
-	private int coinratepct;
-	private int coinpremium;
+	private double coinratepct;
+	private double coinpremium;
 	private int coindeal;
 	
 	public String getMarketname() {
@@ -30,19 +30,19 @@ public class CoinVO {
 	public int getCoinrate() {
 		return coinrate;
 	}
-	public void setCoinratepct(int coinratepct) {
+	public void setCoinratepct(double coinratepct) {
 		this.coinratepct = coinratepct;
 	}
-	public int getCoinratepct() {
+	public double getCoinratepct() {
 		return coinratepct;
 	}
 	public void setCoinrate(int coinrate) {
 		this.coinrate = coinrate;
 	}
-	public int getCoinpremium() {
+	public double getCoinpremium() {
 		return coinpremium;
 	}
-	public void setCoinpremium(int coinpremium) {
+	public void setCoinpremium(double coinpremium) {
 		this.coinpremium = coinpremium;
 	}
 	public int getCoindeal() {
